Observers/nsfw_content_checker.py
from Observers.base_image_checker import BaseImageChecker
from PIL import Image
import numpy as np

class NSFWContentChecker(BaseImageChecker):
    def handle_generated_image(self, image: Image.Image):
        self.last_check_result = self._check_nsfw(image)
        if self.last_check_result:
            print(f"[NSFWContentChecker] Предупреждение: Обнаружен NSFW контент")

    def _check_nsfw(self, image: Image.Image) -> bool:
        # Заглушка для реальной реализации
        return False